# Speed up prober unit tests.