In my local testing, these tests used to take 25s and sometimes more. This reduces the runtime of the tests in the package to ~1.5s.

<!--
Request Prow to automatically lint any go code in this PR:

/lint
-->

## Proposed Changes

In my local testing, these tests used to take 25s and sometimes more. This reduces the runtime of the tests in the package to ~1.5s.

**Release Note**

<!-- Enter your extended release note in the below block. If the PR requires
additional action from users switching to the new release, include the string
"action required". If no release note is required, write "NONE". -->

```release-note
NONE
```

/assign @vagababov